* commit '1d303bd5cdb74a4016d44cd6656ae6ddba3ac03b':
  only process self.tmpPath if the file still exists

+++ mx.py
@@ -9746,15 +9746,16 @@ class SafeFileCreation(object):
         # Windows will complain about tmp being in use by another process
         # when calling os.rename if we don't close the file descriptor.
         os.close(self.tmpFd)
-        if exc_value:
-            # If an error occurred, delete the temp file
-            # instead of renaming it
-            os.remove(self.tmpPath)
-        else:
-            # Correct the permissions on the temporary file which is created with restrictive permissions
-            os.chmod(self.tmpPath, 0o666 & ~currentUmask)
-            # Atomic if self.path does not already exist.
-            os.rename(self.tmpPath, self.path)
+        if exists(self.tmpPath):
+            if exc_value:
+                # If an error occurred, delete the temp file
+                # instead of renaming it
+                os.remove(self.tmpPath)
+            else:
+                # Correct the permissions on the temporary file which is created with restrictive permissions
+                os.chmod(self.tmpPath, 0o666 & ~currentUmask)
+                # Atomic if self.path does not already exist.
+                os.rename(self.tmpPath, self.path)
 
 class Archiver(SafeFileCreation):
     """
